requirements.txt
--extra-index-url https://alpine-wheels.github.io/index

APScheduler==3.7.0
Flask==1.1.2
Markdown==3.3.4
PyJWT==1.7.1
Werkzeug==1.0.1
XlsxWriter==1.3.7
azure-identity==1.5.0
azure-mgmt-compute==19.0.0
azure-mgmt-network==17.1.0
azure-mgmt-subscription==1.0.0
boto3==1.17.22
botocore==1.20.22
elastic-apm[flask]==6.0.0
fort==1.0.1
google-api-python-client==2.0.2
google-auth==1.27.1
oauth2client==4.1.3
pendulum==2.1.2
psycopg2==2.8.6
requests==2.25.1
waitress==1.4.4
whitenoise==5.2.0
